+++ system/platform-core/src/main/java/org/platformlambda/core/models/AsyncMultiInbox.java
@@ -108,9 +108,7 @@ public class AsyncMultiInbox extends InboxBase {
         AsyncMultiInbox holder = (AsyncMultiInbox) inboxes.get(inboxId);
         if (holder != null) {
             float diff = (float) (System.nanoTime() - holder.begin) / EventEmitter.ONE_MILLISECOND;
-            float roundTrip = Float.parseFloat(String.format("%.3f", Math.max(0.0f, diff)));
-            // set roundTrip time and remove some metadata that are not relevant for an RPC response
-            reply.setRoundTrip(roundTrip)
+            reply.setRoundTrip(diff)
                     .removeTag(RPC).setTo(null).setReplyTo(null).setTrace(null, null);
             Map<String, Object> annotations = new HashMap<>();
             // decode trace annotations from reply event
@@ -153,7 +151,7 @@ public class AsyncMultiInbox extends InboxBase {
                         metrics.put("service", correlation.to);
                         metrics.put("from", holder.from);
                         metrics.put("exec_time", reply.getExecutionTime());
-                        metrics.put("round_trip", roundTrip);
+                        metrics.put("round_trip", reply.getRoundTrip());
                         metrics.put("start", start);
                         metrics.put("path", holder.tracePath);
                         payload.put("trace", metrics);